validating interface

+++ project/interfaceModule.py
@@ -1,30 +1,44 @@
 import time
 import zipModule
+import fileModule
 from datetime import datetime
+import re
+import os
 
 sleepTime = 5
 path = "C:\\Users\\felipe\\Desktop\\study\\python\\MultipleFiles\\project\\r_input"
+interface_file_path = "C:\\Users\\felipe\\Desktop\\study\\python\\MultipleFiles\\project\\r_config_interface\\interfaces.txt"
 
-def zip_file_listener(path):
+def interface_listener(path, interface_list):
+    interface_list = fileModule.read_txt_file(interface_file_path)
     while True:
-        print('-->> Checking directory ' + path)
+        print('\n -->> Checking directory ' + path)  
         zipFiles = zipModule.zipFiles(path)
         if not zipFiles:
             dt = datetime.now().strftime("%H:%M:%S")
             print('     Could not find any zip file ' + dt) 
         else:
             for filePath in zipFiles:
-                zipModule.unzip(filePath, path)
-                zipModule.removeFile(filePath)
+                if is_it_a_registered_interface(filePath, interface_list):
+                    zipModule.unzip(filePath, path)
+                    zipModule.removeFile(filePath)
+                else:
+                    print('Error not registered file ' + filePath)
         
         time.sleep(sleepTime)
         
 
 
-def file_name_matches_with_interfaces(file_name, interfaces):
+def is_it_a_registered_interface(file_path, interfaces):
+    file_name = os.path.basename(file_path).split('/')[-1].split('_')[0]
+
     if file_name in interfaces:
         return True
     return False
 
 
-zip_file_listener(path)
+interface_list = fileModule.read_txt_file(interface_file_path)
+if interface_list is None:
+    print('ERROR: Check the interfaces in ' + interface_file_path)
+else:
+    interface_listener(path, interface_list)
